BP5: complete 32-bit round-trip support and build the 32-bit tests
Finishes the 32<->32 BP5 port on top of the type changes in this branch:

- BP5Writer: the md.idx per-flush (pos,size) records (FlushPosSizeInfo,
  and its gather tmp/RecvBuffer) were size_t, written 4-byte on 32-bit
  but read back as uint64_t -> corrupt flush offsets. Make them uint64_t.
  Also fix MDataTotalSize accounting to sizeof(uint64_t) to match the
  uint64_t header entries it counts. No-op on 64-bit (size_t==uint64_t).
- BPBase: default StatsBlockSize = SIZE_MAX. The former 2^50 truncates to
  0 in a 32-bit size_t, giving a divide-by-zero in BP3/BP4 sub-blocking.
- Tests: cast LocalValueDim/JoinedDim to size_t at Dims braced-init and
  EXPECT_EQ sites (they truncate to size_t on 32-bit, so an un-cast use is
  a -Wnarrowing error in braces and a runtime mismatch in comparisons);
  build LargeBlocks 64-bit only (its RequiredDiskSpace constexpr overflows
  a 32-bit size_t).

+++ source/adios2/engine/bp5/BP5Writer.cpp
@@ -313,7 +313,7 @@ uint64_t BP5Writer::WriteMetadata(const std::vector<char> &ContigMetaData,
     uint64_t MetaDataSize = 0;
     std::vector<uint64_t> AttrSizeVector;
 
-    MDataTotalSize += SizeVector.size() * sizeof(size_t);
+    MDataTotalSize += SizeVector.size() * sizeof(uint64_t);
     for (size_t a = 0; a < SizeVector.size(); a++)
     {
         if (a < AttributeBlocks.size())
@@ -2362,12 +2362,12 @@ void BP5Writer::FlushData(const bool isFinal)
 
     if (!isFinal)
     {
-        size_t tmp[2];
+        uint64_t tmp[2];
         // aggregate start pos and data size to rank 0
         tmp[0] = m_StartDataPos;
         tmp[1] = databufsize;
 
-        std::vector<size_t> RecvBuffer;
+        std::vector<uint64_t> RecvBuffer;
         if (m_Comm.Rank() == 0)
         {
             RecvBuffer.resize(m_Comm.Size() * 2);
